+++ apps/kitchen-sink/components/custom/component-previewer/index.tsx
@@ -110,17 +110,17 @@ export const ComponentPreviewer = ({
 
   return (
     <Box className="p-5 min-h-fit rounded-lg m-3 mt-5 border border-outline-100 gap-5 w-full self-center flex-1">
-      <Box className="border-b border-outline-100 pb-2">
+      <Box className={title && `border-b border-outline-100 pb-2`}>
         <Text className="text-typography-900 text-lg font-semibold pb-2">
           {title}
         </Text>
       </Box>
       <Box className="md:flex-row">
-      <Center className="min-h-[100px] md:border-r border-outline-100 flex-1 py-5 w-full">
+      <Center className={`min-h-[100px] ${Object.keys(props).length > 0 ? 'md:border-r' : ''} border-outline-100 flex-1 py-5 w-full`}>
           {children(selectedValues)}
         </Center>
         {Object.keys(props).length > 0 && (
-          <Box className="flex-1 web:items-start web:justify-center w-full web:items-center min-h-fit border-t md:border-t-0">
+          <Box className="flex-1 web:items-start web:justify-center w-full web:items-center min-h-fit border-t border-outline-100 md:border-t-0">
     
             {Object.keys(props).map((key) => (
               <Box