+++ packages/gluestack-ui/src/util/config/next-config-helper.ts
@@ -2,20 +2,13 @@ import * as path from 'path';
 import fg from 'fast-glob';
 import { pathExists, readFile, writeFile } from 'fs-extra';
 import { config } from '../../config';
-import {
-  findDirectory,
-  generateConfig,
-  getFilePath,
-  pathResolver,
-  _currDir,
-} from '.';
+import { findDirectory, getFilePath, pathResolver, _currDir } from '.';
 import {
   RawConfig,
   NextResolvedConfig,
   PROJECT_SHARED_IGNORE,
 } from './config-types';
-import { join, relative } from 'path';
-import { execSync } from 'child_process';
+import { join } from 'path';
 import { ensureFilesPromise } from '..';
 import { commonInitialization } from '../init';
 import os from 'os';
@@ -165,7 +158,7 @@ async function generateConfigNextApp(
     },
   };
 
-  generateConfig(gluestackConfig);
+  // generateConfig(gluestackConfig);
   const resolvedConfig = await resolvedNextJsPaths(resolvedGluestackConfig);
   const filesTobeEnsured = [
     resolvedConfig.app.registry ?? '',