+++ scripts/mappers/starter-kit-next/componentsOperations.ts
@@ -1,15 +1,12 @@
 import path from 'path';
 import {
   processComponentChange,
-  copyUtils,
   MapperConfig,
 } from '../utils/componentOperations';
 
 const mapperConfig: MapperConfig = {
   sourcePath: path.resolve('src/components/ui'),
   destPath: path.resolve('apps/starter-kit-next/components/ui'),
-  utilsSourcePath: path.resolve('src/utils/gluestack-utils'),
-  utilsDestPath: path.resolve('apps/starter-kit-next/utils/gluestack-utils'),
   ignoreFiles: ['docs', 'examples', 'dependencies.json'], // ignore docs, examples and dependencies.json files
 };
 
@@ -17,6 +14,6 @@ export const copyComponent = (component: string, event: string = 'added') => {
   processComponentChange(component, event, mapperConfig);
 };
 
-export const processNonComponentFile = (filePath: string) => {
-  copyUtils(mapperConfig);
-};
+// export const processNonComponentFile = (filePath: string) => {
+//   // Utils are now published as a package, no need to copy them
+// };