+++ packages/gluestack-ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 ## 0.4.9-alpha.0 (2023-07-21)
 
+## 0.7.27
+
+### Patch Changes
+
+- fix: ignored dependencies.json file in init
+
+## 0.7.26
+
+### Patch Changes
+
+- fix: handled empty dependency.json file
+
+## 0.7.25
+
+### Patch Changes
+
+- feat: added support for installing cross dependencies recursively
+
 ## 0.7.24
 
 ### Patch Changes